Implement content negotiation for instances endpoint and add CBD serialization
